test-scaffold/pyproject.toml
[project]
name = "test-scaffold"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
]
